requirements.txt
openai
loguru
pyside6
pyinstaller
pyarmor
pywin32
apscheduler
requests
python-dotenv